Fix pointer cursor on icons in primary button
The cursor in primary buttons is shown as a pointer (if the button is
not disabled) to convey that it can be interacted with. However, if an
icon is included in the button the icon would use the default cursor (as
the icon is typically included through a "span" element), which causes a
weird cursor change between the button icon and the rest of the button.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>